requirements.txt
gspread
google-api-python-client
oauth2client
python-dotenv
libretranslatepy
